Cleanup the usage of drivers
Package consumers now need to establish their own connections.
DB objects can now be added and removed from MultiDB.

This commit greatly simplifies the code and scope of this package.
It breaks backward compaitbilty with version <= 0.1.
The Tx type of this package has been dropped in favor of the sql.Tx type.

Related to #2, this means that Node now implements boil.Beginner.
Closes #16